Address PR review: dedup cliPath into common-utils; make agent-server spawn options caller-controllable
Review feedback from robgruen on the CodeQL fix PR:

1. cliPath.ts duplication: the secure execFileSync where/which helper was duplicated across agentSdk, actionGrammar, and kp. Hoisted the single implementation into @typeagent/common-utils (node-only export); the three packages now re-export from it, so there is one source of truth. Removed common-utils's unused @typeagent/agent-sdk dependency to keep the dep graph acyclic (agent-sdk now depends on common-utils). kp's copy still used the vulnerable execSync pattern, so it inherits the fix too.

2. agent-server spawn behavior: added an optional AgentServerSpawnOptions param threaded through ensureAgentServer -> spawnAgentServer so callers can override detached/windowsHide/stdio for the direct-node spawn paths. Defaults preserve the existing background-server behavior.
