+++ ts/packages/actionGrammar/package.json
@@ -47,6 +47,7 @@
   "dependencies": {
     "@anthropic-ai/claude-agent-sdk": "^0.3.150",
     "@typeagent/action-schema": "workspace:*",
+    "@typeagent/common-utils": "workspace:*",
     "@typeagent/config": "workspace:*",
     "debug": "^4.4.0",
     "dotenv": "^16.4.5",
