+++ ts/packages/kp/package.json
@@ -34,6 +34,7 @@
   "dependencies": {
     "@anthropic-ai/claude-agent-sdk": "^0.3.150",
     "@typeagent/aiclient": "workspace:*",
+    "@typeagent/common-utils": "workspace:*",
     "better-sqlite3": "12.6.2",
     "debug": "^4.4.0",
     "typechat": "^0.1.1"
