+++ ts/packages/utils/commonUtils/src/indexNode.ts
@@ -17,3 +17,5 @@ export {
 export { uint8ArrayToBase64, base64ToUint8Array } from "./base64Node.js";
 
 export { createPromiseWithResolvers } from "./promiseWithResolvers.js";
+
+export { resolveCliOnPath, claudeExecutableOption } from "./cliPath.js";
